packages/platform-core/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "platform-core"
version = "0.1.0"
description = "Shared platform library for agent runtimes — memory, RAG, HITL, reasoning, tools, observability, prompts."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "AEA Platform Team" }
]
keywords = ["agents", "agentic", "ai", "platform", "healthcare"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0,<7.0",
    "pydantic>=2.0,<3.0",
    "requests>=2.32.0,<3.0",
    "fastapi>=0.129.0,<1.0",
    "langgraph>=1.0.8,<2.0",
    "langchain>=1.2.10,<2.0",
    "langchain-openai>=1.1.10,<2.0",
    "boto3>=1.42.0,<2.0",
    "python-dotenv>=1.0.0,<2.0",
    "pyjwt[crypto]>=2.9.0,<3.0",
    "jsonschema>=4.22.0,<5.0",
    "openai>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.6,<1.0",
    "mypy>=1.10,<2.0",
    "types-pyyaml",
    "types-requests",
    "pip-audit>=2.7,<3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
platform_core = ["py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 110
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (covered by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.11"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q --strict-markers"
